requirements.txt
# Bot dependencies
python-telegram-bot>=20.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
